review comments incorporated
