+++ src/styles/components/Wizard.scss
@@ -84,6 +84,9 @@
                     font-size: $font-size-7;
                     font-weight: $font-weight-regular;
                     color: $gray-900;
+                    &:focus {
+                        outline: none !important;
+                    }
                 }
             }
         }
